improved handling of timming differences between clock and mmap
The MMAP_1 signal is noticablylater than the clock due to all the gates it has to go through in the memory map mananger. This caused issues with the trailing edge enable fo the LCD2004 module in that the inverted clock would go high moments before the MMAP_1 signal would go low. This in turn caused a spurious trailing edge event into the Enable input of the LCD2004 module, causing weirdness with the cursor.